+++ phpme/command/phpme_do_generate_method_command.py
@@ -13,12 +13,6 @@ class PhpmeDoGenerateMethodCommand(sublime_plugin.TextCommand):
         # insert uses
         self.view.run_command('phpme_post_use_class', {'namespaces': arranged['uses']})
 
-        # Find the closing brackets. We'll place the method
-        # stubs just before the last closing bracket.
-        closing_brackets = self.view.find_all('}')
-        region = closing_brackets[-1]
-        point = region.end() - 1
-
-        self.view.insert(edit, point, '\n\t' + ('\n\n\t'.join(arranged['methods']))+'\n')
+        self.view.insert(edit, self.view.sel()[0].end(), '\n\n\t'.join(arranged['methods']))
         helper.print_message('Successfully ' + job + ': ' + (', '.join(arranged['arranged'])))